playwright_test/requirements.txt
# Playwright 自动化测试依赖
playwright==1.40.0
pytest==7.4.3
pytest-xdist==3.5.0
pyyaml==6.0.1
allure-pytest==2.13.2
